Tracker: move update_scan into mode.cpp
Consolidating Mode:: methods in mode.cpp
